feat: add hooks for multi-select, pagination, search, sorting, and tabs functionality
- Implemented `useMultiSelect` for managing multiple selections with query state.
- Added `usePagination` for handling pagination logic and state management.
- Created `useSearch` for search functionality with debouncing and validation.
- Developed `useSorting` for sorting data with customizable columns and orders.
- Introduced `useTabs` for managing tab states and navigation.
- Removed unused date and debounce utility functions.
- Added unit tests to verify the functionality and exports of the new hooks.
