--- registry/default/date-range/index.ts
@@ -1,3 +1,0 @@
-export * from './presets'
-export type { DateFormat, DatePreset, UseDateRangeOptions, UseDateRangeResult } from './types'
-export { useDateRange } from './useDateRange'
