--- registry/default/filtering/index.ts
@@ -1,2 +1,0 @@
-export type { UseFiltersOptions, UseFiltersResult, Values } from './types'
-export { useFilters } from './useFilters'
