--- registry/default/index.ts
@@ -1,3 +1,0 @@
-export * from './hooks'
-export * from './types'
-export * from './utils'
