--- registry/default/multi-select/index.ts
@@ -1,2 +1,0 @@
-export type { UseMultiSelectOptions, UseMultiSelectResult } from './types'
-export { useMultiSelect } from './useMultiSelect'
