--- registry/default/pagination/index.ts
@@ -1,2 +1,0 @@
-export type { UsePaginationOptions, UsePaginationResult } from './types'
-export { usePagination } from './usePagination'
