--- registry/default/search/index.ts
@@ -1,2 +1,0 @@
-export type { UseSearchOptions, UseSearchResult } from './types'
-export { useSearch } from './useSearch'
