--- registry/default/sorting/index.ts
@@ -1,2 +1,0 @@
-export type { SortOrder, UseSortingOptions, UseSortingResult } from './types'
-export { useSorting } from './useSorting'
