--- registry/default/tabs/index.ts
@@ -1,2 +1,0 @@
-export type { UseTabsOptions, UseTabsResult } from './types'
-export { useTabs } from './useTabs'
