--- registry/default/utils/date.ts
@@ -1,30 +1,0 @@
-export function startOfDay(date: Date): Date {
-  const newDate = new Date(date)
-  newDate.setHours(0, 0, 0, 0)
-  return newDate
-}
-
-export function endOfDay(date: Date): Date {
-  const newDate = new Date(date)
-  newDate.setHours(23, 59, 59, 999)
-  return newDate
-}
-
-export function addDays(date: Date, days: number): Date {
-  const result = new Date(date)
-  result.setDate(result.getDate() + days)
-  return result
-}
-
-export function startOfMonth(date: Date): Date {
-  return new Date(date.getFullYear(), date.getMonth(), 1)
-}
-
-export function endOfMonth(date: Date): Date {
-  return new Date(date.getFullYear(), date.getMonth() + 1, 0, 23, 59, 59, 999)
-}
-
-export function daysBetween(start: Date, end: Date): number {
-  const msPerDay = 24 * 60 * 60 * 1000
-  return Math.round(Math.abs((end.getTime() - start.getTime()) / msPerDay))
-}
