--- registry/default/utils/index.ts
@@ -1,3 +1,0 @@
-export * from './date'
-export * from './debounce'
-export * from './validation'
