--- registry/default/utils/validation.ts
@@ -1,11 +1,0 @@
-export function clamp(value: number, min: number, max: number): number {
-  return Math.min(Math.max(value, min), max)
-}
-
-export function isValidNumber(value: unknown): value is number {
-  return typeof value === 'number' && !Number.isNaN(value) && Number.isFinite(value)
-}
-
-export function isValidDate(date: unknown): date is Date {
-  return date instanceof Date && !Number.isNaN(date.getTime())
-}
